removing head folder, straight to google
