+++ DESCRIPTION
@@ -4,14 +4,11 @@ Title: R Client for BigQuery Storage API
 Version: 0.0.9999
 Authors@R: c(
   person("Bruno", "Tremblay", role = c("aut", "cre"), email = "openr@neoxone.com"),
-  person("Neal", "Fultz", role = "aut", email = "nfultz@gmail.com"),
   person(family = "Google LLC", role = c("cph", "fnd")))
 Maintainer: Bruno Tremblay <openr@neoxone.com>
 Description: Provides facilities to interact with BigQuery Storage API.
 License: file LICENSE
 Imports:
-    RProtoBuf,
-    R6,
     Rcpp
 Encoding: UTF-8
 LazyData: true
