+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
-export(bqs_client)
 export(grpc_version)
 exportPattern("^[[:alpha:]]+")
 importFrom(Rcpp,evalCpp)
+importFrom(Rcpp,sourceCpp)
 useDynLib(bigrquerystorage)
