+++ R/bigrquerystorage.R
@@ -2,3 +2,7 @@
 #' @importFrom Rcpp evalCpp
 #' @exportPattern "^[[:alpha:]]+"
 NULL
+
+.onUnload <- function (libpath) {
+  library.dynam.unload("bigrquerystorage", libpath)
+}
