+++ cleanup
@@ -1,0 +1,4 @@
+#!/bin/sh
+
+# Remove automatically generated Makevars
+rm -f src/Makevars
