--- man/bqs_client.Rd
@@ -1,49 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/bqs_client.R
-\name{bqs_client}
-\alias{bqs_client}
-\title{Build a client handle}
-\usage{
-bqs_client(verbose = FALSE, credentials, credentials_file)
-}
-\arguments{
-\item{verbose}{Default `FALSE`. When `TRUE`, equivalent to setting
-environment variable `GRPC_VERBOSITY` to `DEBUG` but redirecting
-`stdout` to R console.}
-
-\item{credentials}{A valid access token. (WIP Not implemented)}
-
-\item{credentials_file}{A path to a service account file.}
-}
-\value{
-client handle
-}
-\description{
-Build a client handle
-}
-\details{
-`credentials` and `credentials_file` are mutually exclusive.
-Environment variable `GOOGLE_APPLICATION_CREDENTIALS` will be set with
-`credentials_file` value when provided.
-
-If your application runs inside a Google Cloud environment that has
-a default service account, your application can retrieve the service
-account credentials to call Google Cloud APIs. Such environments
-include Compute Engine, Google Kubernetes Engine, App Engine,
-Cloud Run, and Cloud Functions. We recommend using this strategy
-because it is more convenient and secure than manually passing credentials.
-
-Additionally, we recommend you use Google Cloud Client Libraries for
-your application. Google Cloud Client Libraries use a library called
-Application Default Credentials (ADC) to automatically find your service
-account credentials. ADC looks for service account credentials
-in the following order:
-
-1. If the environment variable GOOGLE_APPLICATION_CREDENTIALS is set,
-ADC uses the service account file that the variable points to.
-2. If the environment variable GOOGLE_APPLICATION_CREDENTIALS isn't
-set, ADC uses the default service account that Compute Engine,
-Google Kubernetes Engine, App Engine, Cloud Run, and Cloud
-Functions provide.
-3. If ADC can't use either of the above credentials, an error occurs.
-}
