+++ src/bqs.cpp
@@ -1,0 +1,68 @@
+// #include <chrono>
+// #include <iostream>
+// #include <memory>
+// #include <random>
+// #include <string>
+// #include <thread>
+
+#include <grpc/grpc.h>
+#include <grpcpp/channel.h>
+#include <grpcpp/client_context.h>
+#include <grpcpp/create_channel.h>
+#include <grpcpp/security/credentials.h>
+#include <grpc/impl/codegen/log.h>
+#include "google/cloud/bigquery/storage/v1/arrow.pb.h"
+#include "google/cloud/bigquery/storage/v1/avro.pb.h"
+#include "google/cloud/bigquery/storage/v1/stream.pb.h"
+#include "google/cloud/bigquery/storage/v1/storage.pb.h"
+#include "google/cloud/bigquery/storage/v1/storage.grpc.pb.h"
+#include <Rcpp.h>
+
+using namespace Rcpp;
+using grpc::Channel;
+using grpc::ClientContext;
+using grpc::ClientReader;
+using grpc::ClientReaderWriter;
+using grpc::ClientWriter;
+using grpc::Status;
+using google::cloud::bigquery::storage::v1::ArrowRecordBatch;
+using google::cloud::bigquery::storage::v1::ArrowSchema;
+using google::cloud::bigquery::storage::v1::AvroRows;
+using google::cloud::bigquery::storage::v1::AvroSchema;
+using google::cloud::bigquery::storage::v1::CreateReadSessionRequest;
+using google::cloud::bigquery::storage::v1::DataFormat;
+using google::cloud::bigquery::storage::v1::ReadRowsRequest;
+using google::cloud::bigquery::storage::v1::ReadRowsResponse;
+using google::cloud::bigquery::storage::v1::ReadSession;
+using google::cloud::bigquery::storage::v1::ReadSession_TableModifiers;
+using google::cloud::bigquery::storage::v1::ReadSession_TableReadOptions;
+using google::cloud::bigquery::storage::v1::ReadStream;
+using google::cloud::bigquery::storage::v1::SplitReadStreamRequest;
+using google::cloud::bigquery::storage::v1::SplitReadStreamResponse;
+using google::cloud::bigquery::storage::v1::StreamStats;
+using google::cloud::bigquery::storage::v1::StreamStats_Progress;
+using google::cloud::bigquery::storage::v1::ThrottleState;
+using google::cloud::bigquery::storage::v1::BigQueryRead;
+
+
+void rgpr_default_log(gpr_log_func_args* args) {
+  Rcout << args->message << std::endl;
+}
+
+//' Check grpc version
+//' @return Version string and what g stands for
+//' @export
+// [[Rcpp::export]]
+CharacterVector grpc_version() {
+  return CharacterVector::create(grpc_version_string(), grpc_g_stands_for());
+}
+
+class BigQueryReadClient {
+public:
+  BigQueryReadClient(std::shared_ptr<Channel> channel)
+    : stub_(BigQueryRead::NewStub(channel)) {
+  }
+private:
+  std::unique_ptr<BigQueryRead::Stub> stub_;
+};
+
